Move fsck out of experimental, and expose fsck library
