+++ storage/posix/files_test.go
@@ -26,7 +26,7 @@ import (
 	"github.com/transparency-dev/tessera"
 	"github.com/transparency-dev/tessera/api"
 	"github.com/transparency-dev/tessera/api/layout"
-	"github.com/transparency-dev/tessera/internal/fsck"
+	"github.com/transparency-dev/tessera/fsck"
 	"golang.org/x/mod/sumdb/note"
 )
